Homework2_DSBD/data_collector/requirements.txt
yfinance
mysql-connector-python
schedule
confluent_kafka
